+++ src/plugins/rtf-to-txt/index.js
@@ -13,7 +13,7 @@ const { v4 } = require("uuid");
  * @author Frazer Smith
  * @description Pre-handler plugin that uses UnRTF to convert Buffer or string of
  * RTF file in `req.body` to TXT and places RTF file in a temporary directory.
- * `req` object is decorated with `rtfToTxtResults` object detailing document
+ * `req` object is decorated with `conversionResults` object detailing document
  * location, contents etc.
  * @param {Function} server - Fastify instance.
  * @param {object} options - Fastify config values.
@@ -26,24 +26,26 @@ const { v4 } = require("uuid");
  */
 async function plugin(server, options) {
 	server.addHook("onRequest", async (req) => {
-		req.rtfToTxtResults = { body: undefined, docLocation: {} };
+		req.conversionResults = { body: undefined };
 	});
 
 	server.addHook("onResponse", (req, res) => {
-		// Remove files from temp directory after response sent
-		const files = glob.sync(
-			`${req.rtfToTxtResults.docLocation.directory}/${req.rtfToTxtResults.docLocation.id}*`
-		);
-		files.forEach((file) => {
-			fs.unlinkSync(file);
-		});
+		if (req.conversionResults.docLocation) {
+			// Remove files from temp directory after response sent
+			const files = glob.sync(
+				`${req.conversionResults.docLocation.directory}/${req.conversionResults.docLocation.id}*`
+			);
+			files.forEach((file) => {
+				fs.unlinkSync(file);
+			});
+		}
 
 		return res;
 	});
 
 	server.addHook("preHandler", async (req, res) => {
 		try {
-			// Define any default settings the middleware should have to get up and running
+			// Define any default settings the plugin should have to get up and running
 			const config = {
 				binPath: undefined,
 				rtfToTxtOptions: {
@@ -73,12 +75,12 @@ async function plugin(server, options) {
 			 * Windows-1252 to UTF-8 results with HTML equivalents.
 			 * Refer to https://www.i18nqa.com/debug/utf8-debug.html for more info.
 			 */
-			req.rtfToTxtResults.body = await unrtf.convert(
+			req.conversionResults.body = await unrtf.convert(
 				tempFile,
 				config.rtfToTxtOptions
 			);
 
-			req.rtfToTxtResults.docLocation = {
+			req.conversionResults.docLocation = {
 				directory: config.tempDirectory,
 				rtf: tempFile,
 				id,
